fix: resolve markdown table rendering and documentation display issues
- Create shared DocumentationViewer and MarkdownRenderer components in ui-components
- Add remark-gfm plugin for GitHub Flavored Markdown table support
- Fix missing margins and scrolling in GoPCA documentation viewer
- Update Tailwind configs to include ui-components path for prose styles
- Clean up duplicate dependencies following DRY principle
- Move markdown libs from peerDependencies to regular dependencies
- Remove unused markdown dependencies from app package.json files

This consolidates documentation rendering into shared components,
eliminating 300+ lines of duplicate code while fixing the markdown
table rendering issue reported in #435.

Fixes #435
