+++ cmd/gocsv/frontend/package.json
@@ -17,13 +17,9 @@
     "ag-grid-react": "^31.0.0",
     "file-saver": "^2.0.5",
     "html-to-image": "^1.11.13",
-    "katex": "^0.16.22",
     "lucide-react": "^0.539.0",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-markdown": "^10.1.0",
-    "rehype-katex": "^7.0.1",
-    "remark-math": "^6.0.0"
+    "react-dom": "^18.2.0"
   },
   "devDependencies": {
     "@types/file-saver": "^2.0.7",
