+++ cmd/gocsv/frontend/package.json.md5
@@ -1,1 +1,1 @@
-bd58044aaac7e4122941f195151fcb68
+473cdd2c5157a8ec87f8cbd368028468
