+++ cmd/gopca-desktop/frontend/package.json
@@ -16,13 +16,9 @@
     "@types/plotly.js": "^3.0.3",
     "file-saver": "^2.0.5",
     "html-to-image": "^1.11.13",
-    "katex": "^0.16.22",
     "lucide-react": "^0.539.0",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-markdown": "^10.1.0",
-    "rehype-katex": "^7.0.1",
-    "remark-math": "^6.0.0"
+    "react-dom": "^18.2.0"
   },
   "devDependencies": {
     "@types/file-saver": "^2.0.7",
