+++ cmd/gopca-desktop/frontend/package.json.md5
@@ -1,1 +1,1 @@
-eb764ae3dfdc410463250ec631018a88
+334bd51379f28009a9c4ee7d237caafb
