+++ cmd/gopca-desktop/frontend/tailwind.config.js
@@ -10,6 +10,7 @@ export default {
   content: [
     "./index.html",
     "./src/**/*.{js,ts,jsx,tsx}",
+    "../../../packages/ui-components/src/**/*.{js,ts,jsx,tsx}",
   ],
   theme: {
     extend: {},
