+++ packages/ui-components/src/index.ts
@@ -15,6 +15,8 @@ export { InputDialog } from './components/InputDialog';
 export { SkipLinks } from './components/SkipLinks';
 export { KeyboardHelp } from './components/KeyboardHelp';
 export { CustomSelect } from './components/CustomSelect';
+export { DocumentationViewer } from './components/DocumentationViewer';
+export { MarkdownRenderer } from './components/MarkdownRenderer';
 
 // Component Types
 export type { ExportButtonProps, ExportConfig, ExportFormat } from './components/ExportButton';
@@ -26,6 +28,8 @@ export type { InputDialogProps } from './components/InputDialog';
 export type { SkipLinksProps, SkipLink } from './components/SkipLinks';
 export type { KeyboardHelpProps } from './components/KeyboardHelp';
 export type { SelectOption } from './components/CustomSelect';
+export type { DocumentationViewerProps } from './components/DocumentationViewer';
+export type { MarkdownRendererProps } from './components/MarkdownRenderer';
 
 // Contexts
 export { ThemeProvider, useTheme } from './contexts/ThemeContext';
